+++ ServerDevcommands/Commands/Bind.cs
@@ -44,10 +44,11 @@ public class BindCommand
     {
       if (index == 0 && subIndex == 0) return ParameterInfo.KeyCodes;
       if (index == 0 && subIndex == 1) return ParameterInfo.KeyCodesWithNegative;
-      return ParameterInfo.Command(index - 1);
+      return ParameterInfo.None;
     }, new() {
       { "keys", (int index) => ParameterInfo.KeyCodesWithNegative }
     });
+    AutoComplete.Offsets["bind"] = 1;
 
     new Terminal.ConsoleCommand("rebind", "[keycode,modifier1,modifier2,...] [command] [parameters] - Binds a key (with modifier keys) to a command.", (args) =>
     {